+++ extract_msg/properties/named.py
@@ -23,7 +23,8 @@ from typing import (
 
 from .. import constants
 from ..enums import NamedPropertyType
-from ..utils import bytesToGuid, divide, msgPathToString
+from ..exceptions import InvalidPropertyIdError
+from ..utils import bytesToGuid, divide, msgPathToString, verifyPropertyId
 from compressed_rtf.crc32 import crc32
 
 
@@ -55,7 +56,8 @@ class Named:
         self.namesStream = self.getStream('__substg1.0_00040102') or b''
 
         self.__propertiesDict: Dict[Tuple[str, str], NamedPropertyBase] = {}
-        self.__properties: List[NamedPropertyBase] = []
+
+        self.__streamIDDict: Dict[str, Tuple[str, str]] = {}
 
         # Check that we even have any entries. If there are none, nothing to do.
         if entryStream:
@@ -73,14 +75,15 @@ class Named:
                 entry['guid'] = guids[entry['guid_index']]
                 entries.append(entry)
 
-            self.entries = entries
+            properties: List[NamedPropertyBase] = []
 
             for entry in entries:
-                self.__properties.append(StringNamedProperty(entry, self.__getName(entry['id'])) if entry['pkind'] == NamedPropertyType.STRING_NAMED else NumericalNamedProperty(entry))
+                properties.append(StringNamedProperty(entry, self.__getName(entry['id'])) if entry['pkind'] == NamedPropertyType.STRING_NAMED else NumericalNamedProperty(entry))
 
-            for property in self.__properties:
-                name = property.name if isinstance(property, StringNamedProperty) else property.propertyID
-                self.__propertiesDict[(name, property.guid)] = property
+            for property in properties:
+                id_ = property.identifier
+                self.__propertiesDict[id_] = property
+                self.__streamIDDict[property.propertyStreamID] = id_
 
     def __contains__(self, key) -> bool:
         return key in self.__propertiesDict
@@ -152,6 +155,38 @@ class Named:
             return self[propertyName]
         except KeyError:
             return default
+
+    def getPropNameByStreamID(self, streamID: Union[int, str]) -> Optional[Tuple[str, str]]:
+        """
+        Gets the name of a property (as a key for the internal dict) that is stored in the specified stream.
+
+        Useful for determining if a stream/property stream entry is a named
+        property.
+
+        :param streamID: A 4 hex character identifier that will be checked. May
+            also be an integer that can convert to 4 hex characters.
+
+        :returns: The name, if the stream is a named property, otherwise
+            ``None``.
+
+        :raises InvalidPropertyIdError: The Stream ID is invalid.
+        :raises TypeError: The Stream ID is not a valid type.
+        """
+        if isinstance(streamID, int):
+            if streamID < 0x8000:
+                # Definitely doesn't refer to a name property, so just return
+                # None.
+                return None
+            if streamID > 0xFFFF:
+                raise InvalidPropertyIdError('Stream ID is out of range.')
+            streamID = f'{streamID:04X}'
+        elif isinstance(streamID, str):
+            verifyPropertyId(streamID)
+        else:
+            raise TypeError(':param streamID: MUST be an int or str.')
+
+        return self.__streamIDDict.get(streamID)
+
 
     def getStream(self, filename: constants.MSG_PATH) -> Optional[bytes]:
         """
@@ -278,6 +313,13 @@ class NamedPropertyBase(abc.ABC):
         return self.__guidIndex
 
     @property
+    @abc.abstractmethod
+    def identifier(self) -> Tuple[str, str]:
+        """
+        An identifier that can be used to full identify the property.
+        """
+
+    @property
     def namedPropertyID(self) -> int:
         """
         The named property id.
@@ -345,6 +387,10 @@ class StringNamedProperty(NamedPropertyBase):
             self.__streamID = 0x1000 + (crc32(name.encode('utf-16-le')) ^ (self.guidIndex << 1 | 1)) % 0x1F
 
     @property
+    def identifier(self) -> Tuple[str, str]:
+        return [self.name, self.guid]
+
+    @property
     def name(self) -> str:
         """
         The name of the property.
@@ -373,6 +419,10 @@ class NumericalNamedProperty(NamedPropertyBase):
         super().__init__(entry)
         self.__propertyID = f'{entry["id"]:04X}'
         self.__streamID = 0x1000 + (entry['id'] ^ (self.guidIndex << 1)) % 0x1F
+
+    @property
+    def identifier(self) -> Tuple[str, str]:
+        return [self.propertyID, self.guid]
 
     @property
     def propertyID(self) -> str: